refactor: reusable function
